feat: scrollable state table
